# server/swagger.yaml
...
openapi: 3.0.0
info:
  title: ShriDevPAthshalA
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /classes/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Classes
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Classes
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classes"
    post:
      summary: CRUD ACTION update
      tags:
        - Classes
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Classes"
    
  #-----  Exam  -----
  /exam/{ID}/validate:
    post:
      summary: THIS API USED TO VALIDATE THE EXAM
      security: []   # No security
      tags:
        - Exam
      parameters:
        - name: iD
          in: path
          description: iD OF EXAM
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  /exam/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Exam
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Exam
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    post:
      summary: CRUD ACTION update
      tags:
        - Exam
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Exam"
  /exam/findBy_Class/{key}:
    get:
      summary: CRUD ACTION findBy_Class
      tags:
        - Exam
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
  /exam/findBy_Student/{key}:
    get:
      summary: CRUD ACTION findBy_Student
      tags:
        - Exam
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
  /exam/findBy_Teacher/{key}:
    get:
      summary: CRUD ACTION findBy_Teacher
      tags:
        - Exam
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    
  #-----  Student  -----
...
  /student/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Student
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Student
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    post:
      summary: CRUD ACTION update
      tags:
        - Student
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
  /student/findBy_Classes/{key}:
    get:
      summary: CRUD ACTION findBy_Classes
      tags:
        - Student
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    
  #-----  Teacher  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Classes:
      properties:
        _id:
          type: string
        Name:
          type: string
      required:
        -  _id
        - Name
    Exam:
      properties:
        _id:
          type: string
        Name:
          type: string
        Score:
          type: number
        vALID:
          type: boolean
      required:
        -  _id
    Student:
      properties:
        _id:
          type: string
        DOB:
          type: integer
        FatherName:
          type: string
        MobileNumber:
          type: number
        MotherName:
          type: string
        Name:
          type: string
        Surname:
          type: string
      required:
        -  _id
        - DOB
        - FatherName
        - MobileNumber
        - MotherName
        - Name
...